+++ src/Entities/Shared/KoreaLocalPaymentMethodType.php
@@ -1,0 +1,76 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * |------
+ * | ! Generated code !
+ * | Altering this code will result in changes being overwritten |
+ * |-------------------------------------------------------------|.
+ */
+
+namespace Paddle\SDK\Entities\Shared;
+
+use Paddle\SDK\PaddleEnum;
+
+/**
+ * @method static KoreaLocalPaymentMethodType BC()
+ * @method static KoreaLocalPaymentMethodType Citi()
+ * @method static KoreaLocalPaymentMethodType Hana()
+ * @method static KoreaLocalPaymentMethodType Hyundai()
+ * @method static KoreaLocalPaymentMethodType Jeju()
+ * @method static KoreaLocalPaymentMethodType Jeonbuk()
+ * @method static KoreaLocalPaymentMethodType KakaoBank()
+ * @method static KoreaLocalPaymentMethodType KakaoPay()
+ * @method static KoreaLocalPaymentMethodType KBank()
+ * @method static KoreaLocalPaymentMethodType KDBBank()
+ * @method static KoreaLocalPaymentMethodType Kookmin()
+ * @method static KoreaLocalPaymentMethodType Kwangju()
+ * @method static KoreaLocalPaymentMethodType Lotte()
+ * @method static KoreaLocalPaymentMethodType Mg()
+ * @method static KoreaLocalPaymentMethodType NaverPayCard()
+ * @method static KoreaLocalPaymentMethodType NaverPayPoint()
+ * @method static KoreaLocalPaymentMethodType NH()
+ * @method static KoreaLocalPaymentMethodType PayCo()
+ * @method static KoreaLocalPaymentMethodType Post()
+ * @method static KoreaLocalPaymentMethodType Samsung()
+ * @method static KoreaLocalPaymentMethodType SamsungPay()
+ * @method static KoreaLocalPaymentMethodType SavingsBank()
+ * @method static KoreaLocalPaymentMethodType Shinhan()
+ * @method static KoreaLocalPaymentMethodType Shinhyup()
+ * @method static KoreaLocalPaymentMethodType Suhyup()
+ * @method static KoreaLocalPaymentMethodType TossBank()
+ * @method static KoreaLocalPaymentMethodType Unknown()
+ * @method static KoreaLocalPaymentMethodType Woori()
+ */
+final class KoreaLocalPaymentMethodType extends PaddleEnum
+{
+    private const BC = 'bc';
+    private const Citi = 'citi';
+    private const Hana = 'hana';
+    private const Hyundai = 'hyundai';
+    private const Jeju = 'jeju';
+    private const Jeonbuk = 'jeonbuk';
+    private const KakaoBank = 'kakaobank';
+    private const KakaoPay = 'kakaopay';
+    private const KBank = 'kbank';
+    private const KDBBank = 'kdbbank';
+    private const Kookmin = 'kookmin';
+    private const Kwangju = 'kwangju';
+    private const Lotte = 'lotte';
+    private const MG = 'mg';
+    private const NaverPayCard = 'naverpaycard';
+    private const NaverPayPoint = 'naverpaypoint';
+    private const NH = 'nh';
+    private const PayCo = 'payco';
+    private const Post = 'post';
+    private const Samsung = 'samsung';
+    private const SamsungPay = 'samsungpay';
+    private const SavingsBank = 'savingsbank';
+    private const Shinhan = 'shinhan';
+    private const Shinhyup = 'shinhyup';
+    private const Suhyup = 'suhyup';
+    private const TossBank = 'tossbank';
+    private const Unknown = 'unknown';
+    private const Woori = 'woori';
+}